+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add `A.Vector.reduce/2` and `A.Vector.reduce/3`
 * Add `A.Vector.split/2` and `A.Vector.split_with/2`
+* Add `A.Vector.frequencies/1` and `A.Vector.frequencies_by/2`
 * Improve performance for  `A.Vector.min/1` and `A.Vector.max/1`
 
 ## v0.4.6 (2021-02-10)